Full username functionality
Logged in page now is able to use the token to request the username (in the form of email) from the server and display it on the home page. The ''Log In' button on the home page does stretch to fit the length of the username, which causes visual errors with longer names.